refactor: move Apollo Federation v2 code under ee/middleware/graphql_federation
Mirror the dispatcher pattern Tyk Streams, upstreambasicauth, and
upstreamoauth use: real EE code lives under `ee/` (covered by
`ee/LICENSE-EE.md`), and a thin gateway-side dispatcher plugs it into
the engine. Build tags alone aren't enough — the directory boundary IS
the license boundary, and PR #8193's federation v2 work was previously
spread across `apidef/` and `gateway/` with only build-tag gating.

Layering:
- New `ee/middleware/graphql_federation/` package (CE-untagged, whole
  directory is EE-only) holds the real implementation: schema
  augmentation (`augment.go`), `_entities` plan + REST/GraphQL entity
  resolvers (`entities_datasource.go`, `entities_datasource_graphql.go`
  with operation-override / federation-passthrough / generated-lookup
  strategies), `_service { sdl }` SDL synthesis with `@link`
  auto-prepend and orphan-Query-field stripping (`service_sdl.go`), and
  the `enginev3.FederationProvider` impl (`provider.go`).
- New `apidef/adapter/gqlengineadapter/enginev3/federation_provider.go`
  defines the CE-visible `FederationProvider` interface plus a no-op
  fallback. The adapter's `EngineConfigV3()` call sites in
  `adapter_udg.go` and `adapter_proxy_only.go` reach the registered
  provider for federation hooks; CE binaries fall through to the no-op.
- `gateway/mw_graphql_federation_ee.go` shrinks from 47 lines of inline
  `BuildFederationSchema` glue to a 29-line dispatcher: an `init()`
  that registers `graphql_federation.NewProvider()` plus a thin method
  body that delegates to the registered provider. Mirrors
  `mw_upstream_basic_auth_ee.go` exactly.
- `gateway/mw_graphql_federation.go` (CE stub) unchanged.

Subgraph mode preserved in CE: the proxy-mode `UpstreamSchema` setter
in `adapter_proxy_only.go` and its `proxySchemaIsFederated` helper
stay CE-visible, since `GraphQLExecutionModeSubgraph` (hand-written
federation SDL) is a longstanding CE feature that needs the same
upstream-schema wiring as proxy-mode federation passthrough. The
`@key` AST helper is duplicated rather than imported from `ee/` to
keep the layering rule clean.

Tests:
- 11 federation unit tests (TestServiceSDL_*, TestPickLookupField_*,
  TestUDGFederation_Subscription_PlanShape,
  TestUniversalDataGraph_FederationSubgraph,
  TestEntitiesDataSourcePlanner_NoSubscriptions) moved out of the
  adapter package into `ee/middleware/graphql_federation/`. They run
  under the EE build only via `package graphql_federation` directly.
- 21 federation integration tests in `gateway/mw_graphql_federation_test.go`
  stay in `package gateway` with the `ee || dev` build tag (Streams
  uses the same shape with `mw_streaming_test.go`).
- Test-only adapter helpers (`NewUDGForTest`,
  `NewUDGForTestWithSubscriptionFactory`,
  `SetSubscriptionClientFactoryForTest`) added in
  `apidef/adapter/gqlengineadapter/enginev3/test_helpers.go` so the
  EE-package tests exercise the engine without depending on the
  gateway.

Layering check (`grep -rn 'tyk/ee/' apidef/ internal/ header/`)
returns empty. `gofmt -l` and `go vet` clean for both CE and EE
builds. All 32 federation tests (11 in `ee/middleware/graphql_federation/`,
21 in `gateway/`) pass under `-tags=ee`; CE build compiles federation
behavior dormant and the CE federation stub test passes.
