+++ apidef/adapter/gqlengineadapter/enginev3/adapter_proxy_only.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 	"strings"
 
+	"github.com/TykTechnologies/graphql-go-tools/v2/pkg/astparser"
 	graphqldatasource "github.com/TykTechnologies/graphql-go-tools/v2/pkg/engine/datasource/graphql_datasource"
 	"github.com/TykTechnologies/graphql-go-tools/v2/pkg/graphql"
 	"github.com/TykTechnologies/tyk/apidef"
@@ -57,12 +58,15 @@ func (p *ProxyOnly) EngineConfigV3() (*graphql.EngineV2Configuration, error) {
 	// (proxy-mode federation passthrough or subgraph mode), the proxy data
 	// source needs `UpstreamSchema` set to the full federation-aware SDL —
 	// otherwise the planner's abstract-selection rewriter cannot resolve
-	// `_Entity` and `_entities` queries fail at planning time. For plain
-	// (non-federation) proxy schemas we leave the data source untouched to
-	// preserve existing behaviour.
+	// `_Entity` and `_entities` queries fail at planning time. The hook lives
+	// in CE because subgraph mode (`GraphQLExecutionModeSubgraph`) is a
+	// longstanding CE feature — its hand-written federation SDL needs the
+	// same upstream-schema wiring. For plain (non-federation) proxy schemas
+	// we leave the data source untouched.
 	if err == nil && p.Schema != nil && proxySchemaIsFederated(p.ApiDefinition.GraphQL.Schema, p.Schema) {
 		augmentedSDL := string(p.Schema.Document())
 		dataSources := v2Config.DataSources()
+		modified := false
 		for i, ds := range dataSources {
 			var cfg graphqldatasource.Configuration
 			if err := json.Unmarshal(ds.Custom, &cfg); err != nil {
@@ -73,12 +77,74 @@ func (p *ProxyOnly) EngineConfigV3() (*graphql.EngineV2Configuration, error) {
 			}
 			cfg.UpstreamSchema = augmentedSDL
 			dataSources[i].Custom = graphqldatasource.ConfigJson(cfg)
+			modified = true
 		}
-		v2Config.SetDataSources(dataSources)
+		if modified {
+			v2Config.SetDataSources(dataSources)
+		}
 	}
 
 	v2Config.EnableSingleFlight(false)
 	return &v2Config, err
+}
+
+// proxySchemaIsFederated reports whether the proxy mode is fronting an Apollo
+// Federation subgraph. Two signals trigger this:
+//  1. The customer's original SDL declares any `@key` directive (federation
+//     passthrough — gateway middleware then augments the schema before reaching
+//     here).
+//  2. The parsed schema already exposes `@key` types after normalization
+//     (subgraph mode, where customers hand-write the federation schema
+//     themselves).
+//
+// This helper lives in CE because subgraph mode is a CE feature; the EE-only
+// federation v2 work (UDG entity resolvers, schema augmentation, `_service`
+// SDL synthesis) lives under `ee/middleware/graphql_federation/`.
+func proxySchemaIsFederated(customerSDL string, parsed *graphql.Schema) bool {
+	if customerSDL != "" && schemaContainsKeyDirective(customerSDL) {
+		return true
+	}
+	if parsed != nil {
+		printed := string(parsed.Document())
+		if schemaContainsKeyDirective(printed) {
+			return true
+		}
+	}
+	return false
+}
+
+// schemaContainsKeyDirective parses the SDL and reports whether any object
+// type definition or extension carries the `@key` directive. Mirrors the same
+// helper in `ee/middleware/graphql_federation` (kept duplicated to avoid CE
+// importing from `ee/`).
+func schemaContainsKeyDirective(sdl string) bool {
+	doc, report := astparser.ParseGraphqlDocumentString(sdl)
+	if report.HasErrors() {
+		return false
+	}
+	for i := range doc.ObjectTypeDefinitions {
+		def := doc.ObjectTypeDefinitions[i]
+		if !def.HasDirectives {
+			continue
+		}
+		for _, dRef := range def.Directives.Refs {
+			if doc.DirectiveNameString(dRef) == "key" {
+				return true
+			}
+		}
+	}
+	for i := range doc.ObjectTypeExtensions {
+		ext := doc.ObjectTypeExtensions[i]
+		if !ext.HasDirectives {
+			continue
+		}
+		for _, dRef := range ext.Directives.Refs {
+			if doc.DirectiveNameString(dRef) == "key" {
+				return true
+			}
+		}
+	}
+	return false
 }
 
 func parseSchema(schemaAsString string) (parsedSchema *graphql.Schema, err error) {
@@ -114,30 +180,6 @@ func graphqlSubscriptionType(subscriptionType apidef.SubscriptionType) graphql.S
 		// Preview-labeled, so this is not a back-compat concern.
 		return graphql.SubscriptionTypeGraphQLTransportWS
 	}
-}
-
-// proxySchemaIsFederated reports whether the proxy mode is fronting an Apollo
-// Federation subgraph. Two signals trigger this:
-//  1. The customer's original SDL declares any `@key` directive (federation
-//     passthrough — gateway middleware then augments the schema before reaching
-//     here).
-//  2. The parsed schema already exposes `_entities` on Query (subgraph mode,
-//     where customers hand-write the federation schema themselves).
-func proxySchemaIsFederated(customerSDL string, parsed *graphql.Schema) bool {
-	if customerSDL != "" {
-		entityTypes, err := keyedEntityTypes(customerSDL)
-		if err == nil && len(entityTypes) > 0 {
-			return true
-		}
-	}
-	if parsed != nil {
-		printed := string(parsed.Document())
-		entityTypes, err := keyedEntityTypes(printed)
-		if err == nil && len(entityTypes) > 0 {
-			return true
-		}
-	}
-	return false
 }
 
 func subscriptionClientFactoryOrDefault(providedSubscriptionClientFactory graphqldatasource.GraphQLSubscriptionClientFactory) graphqldatasource.GraphQLSubscriptionClientFactory {
