+++ gateway/mw_graphql_federation_ee.go
@@ -3,45 +3,27 @@
 package gateway
 
 import (
-	"github.com/TykTechnologies/graphql-go-tools/v2/pkg/federation"
-
 	"github.com/TykTechnologies/tyk/apidef"
+	"github.com/TykTechnologies/tyk/apidef/adapter/gqlengineadapter/enginev3"
+	graphql_federation "github.com/TykTechnologies/tyk/ee/middleware/graphql_federation"
 )
 
+// init registers the EE Apollo Federation v2 provider with the engine V3
+// adapter. The CE build of `mw_graphql_federation.go` registers nothing, so
+// the adapter falls back to its no-op provider — federation behavior is
+// dormant in CE binaries. Mirrors the dispatcher pattern used by streams,
+// upstreambasicauth, and upstreamoauth: the actual implementation lives
+// under `ee/middleware/graphql_federation/` (covered by `ee/LICENSE-EE.md`)
+// and the gateway plugs it in here.
+func init() {
+	enginev3.RegisterFederationProvider(graphql_federation.NewProvider())
+}
+
 // augmentFederationSchema injects the Apollo Federation v2 extensions
-// (`_Entity` union, `_entities`, `_service`) into the customer's SDL when the
-// API is configured for federation. This is the EE-only implementation; the
-// CE stub in mw_graphql_federation.go returns the SDL unchanged and warns
-// when `@key` directives are present.
-//
-// UDG (`ExecutionEngine`) composes upstreams locally and always augments.
-// Proxy-only (`ProxyOnly`) forwards `_entities` queries to an upstream
-// subgraph or router, but Tyk still validates the operation against its
-// known schema first; in that case we auto-detect federation by scanning for
-// `@key` directives — no config flag.
+// (`_Entity` union, `_entities`, `_service`) into the customer's SDL. The
+// real logic lives in `ee/middleware/graphql_federation`; this gateway hook
+// just reaches the registered provider so the call site in `mw_graphql.go`
+// is build-tag agnostic.
 func (m *GraphQLMiddleware) augmentFederationSchema(schemaStr string, executionMode apidef.GraphQLExecutionMode) (string, error) {
-	switch executionMode {
-	case apidef.GraphQLExecutionModeExecutionEngine:
-		augmented, err := federation.BuildFederationSchema(schemaStr, schemaStr)
-		if err != nil {
-			return schemaStr, err
-		}
-		return augmented, nil
-	case apidef.GraphQLExecutionModeProxyOnly:
-		// Apollo Federation v2 proxy-mode passthrough: when the customer's
-		// SDL declares `@key` types, Tyk forwards `_entities` queries to the
-		// upstream subgraph or router. Tyk still validates the operation
-		// against its known schema first, so we must inject the federation
-		// extensions (`_Entity` union, `_entities`, `_service`) here.
-		// Auto-detect from `@key` — no config flag.
-		if !schemaHasKeyDirective(schemaStr) {
-			return schemaStr, nil
-		}
-		augmented, err := federation.BuildFederationSchema(schemaStr, schemaStr)
-		if err != nil {
-			return schemaStr, err
-		}
-		return augmented, nil
-	}
-	return schemaStr, nil
+	return enginev3.GetFederationProvider().AugmentSchema(schemaStr, executionMode)
 }
